Rename contants.ts and update imports
