+++ src/app/(auth)/login/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 
-import { APP_NAME } from "@/lib/contants";
+import { APP_NAME } from "@/lib/constants";
 
 import { LoginForm } from "@/features/auth";
