+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { APP_DESC, APP_NAME } from "@/lib/contants";
+import { APP_DESC, APP_NAME } from "@/lib/constants";
 
 import { ThemeProvider } from "@wrksz/themes/next";
